More enemies, shooting
Changed the enemies so they have varying speeds and health levels. Made it so the grass texture does not block out the characters. Made a spawner that automatically spawns in enemies on a timer. Added the shooting of projectiles with right click. Modified the enemy AI script so the enemies automatically lock on to the player.